# .trae/documents/openapi-specifications.yaml
openapi: 3.0.3
info:
  title: RomAPI - Écosystème d'API Commercial Cameroun
  description: |
    API complète pour l'écosystème RomAPI qui centralise les ressources et services locaux du Cameroun.
    
    ## Authentification
    
    L'API utilise deux méthodes d'authentification :
    - **API Key** : Pour les intégrations tierces (header `X-API-Key`)
    - **Bearer Token** : Pour les utilisateurs authentifiés (header `Authorization: Bearer <token>`)
    
    ## Rate Limiting
    
    Les limites de taux dépendent de votre plan :
    - **Free** : 1,000 requêtes/mois
    - **Pro** : 50,000 requêtes/mois  
    - **Enterprise** : 500,000 requêtes/mois
    
    ## Formats Locaux
    
    - **Téléphone** : Format international (+237XXXXXXXXX) ou local (6XXXXXXXX)
    - **Adresse** : Formats camerounais avec régions et départements
    - **Fuseau horaire** : Africa/Douala (UTC+1)
    
  version: 1.0.0
  contact:
    name: RomAPI Support
    email: support@romapi.cm
    url: https://romapi.cm
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.romapi.cm/v1
    description: Production server
  - url: https://staging-api.romapi.cm/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Development server
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Connexion utilisateur
      description: Authentifie un utilisateur et retourne un token JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "romuald@example.com"
                password:
                  type: string
                  format: password
                  example: "motdepasse123"
      responses:
        '200':
          description: Connexion réussie
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  refresh_token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Inscription utilisateur
      description: Crée un nouveau compte utilisateur
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
              properties:
                name:
                  type: string
                  example: "Romuald Kamga"
                email:
                  type: string
                  format: email
                  example: "romuald@example.com"
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: "motdepasse123"
                role:
                  type: string
                  enum: [user, business]
                  default: user
                  example: "business"
      responses:
        '201':
          description: Inscription réussie
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Compte créé avec succès"
                  user:
                    $ref: '#/components/schemas/User'
        '409':
          description: Email déjà utilisé
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /api-keys:
    get:
      tags:
        - API Keys
      summary: Liste des clés API
      description: Récupère les clés API de l'utilisateur
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Liste des clés API
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApiKey'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - API Keys
      summary: Créer une clé API
      description: Génère une nouvelle clé API
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: "Mon App Mobile"
                permissions:
                  type: array
                  items:
                    type: string
                    enum: [read, write, analytics]
                  example: ["read"]
      responses:
        '201':
          description: Clé API créée
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Clé API créée avec succès"
                  data:
                    allOf:
                      - $ref: '#/components/schemas/ApiKey'
                      - type: object
                        properties:
                          key:
                            type: string
                            description: "Clé complète (visible une seule fois)"
                            example: "romapi_live_1234567890abcdef"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api-keys/{id}:
    delete:
      tags:
        - API Keys
      summary: Supprimer une clé API
      description: Révoque une clé API
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Clé API supprimée
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Clé API supprimée avec succès"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # Health Check
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Clé API pour l'authentification. Format: `romapi_live_xxxxxxxxxxxxxxxx`
        
        Obtenez votre clé API depuis votre dashboard utilisateur.
    
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtenu via l'endpoint `/auth/login`.
        
        Format: `Authorization: Bearer <token>`
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          example: "Romuald Kamga"
        email:
          type: string
          format: email
          example: "romuald@example.com"
        role:
          type: string
          enum: [user, business, admin]
          example: "business"
        plan:
          type: string
          enum: [free, pro, enterprise]
          example: "pro"
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
...
    ApiKey:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "Mon App Mobile"
        key_prefix:
          type: string
          example: "romapi_live_1234"
        permissions:
          type: array
          items:
            type: string
            enum: [read, write, analytics]
        rate_limit:
          type: integer
          example: 1000
        is_active:
          type: boolean
          example: true
        last_used_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Validation failed"
        message:
          type: string
          example: "Les données fournies ne sont pas valides"
        details:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          example: "/v1/businesses"
  responses:
...
    Unauthorized:
      description: Non autorisé - Token manquant ou invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Unauthorized"
            message: "Token d'authentification requis"
            timestamp: "2024-01-15T10:30:00Z"
            path: "/v1/businesses"
...
    NotFound:
      description: Ressource non trouvée
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Not Found"
            message: "Entreprise non trouvée"
            timestamp: "2024-01-15T10:30:00Z"
            path: "/v1/businesses/123"
    ValidationError:
      description: Erreur de validation des données
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  details:
                    type: object
                    properties:
                      fields:
                        type: object
                        additionalProperties:
                          type: array
                          items:
                            type: string
          example:
            error: "Validation Error"
            message: "Les données fournies ne sont pas valides"
            details:
              fields:
                name: ["Le nom est requis"]
                email: ["Format d'email invalide"]
            timestamp: "2024-01-15T10:30:00Z"
            path: "/v1/businesses"
...